requirements.txt
flask
flask-cors
pytesseract
opencv-python
sentence-transformers
scikit-learn
numpy
Pillow
pandas
joblib
PyYAML
matplotlib
